Add some overrides needed. Fix some stuff
